Split out TableWriterHideFieldsTest
